+++ DepositCalculator/Models/ICurrency.cs
@@ -4,5 +4,5 @@ public interface ICurrency
 {
 	string Name { get; set; }
 
-	decimal AnnualInterestRate { get; set; }
+	double AnnualInterestRate { get; set; }
 }
